Refined Server Sync, some gui changes
